extracted SourceDescriptorDetails
