+++ src/renderer/components/BasemapManagement.js
@@ -2,15 +2,6 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { makeStyles } from '@material-ui/core/styles'
 import BackToMapIcon from '@material-ui/icons/ExitToApp'
-import {
-  Button,
-  Stepper, Step, StepLabel, StepContent, Typography
-} from '@material-ui/core'
-
-import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos'
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos'
-import CloseIcon from '@material-ui/icons/Close'
-import SaveIcon from '@material-ui/icons/Save'
 
 import basemap, { setBasemap, clearBasemap } from '../map/basemap'
 import * as ol from 'ol'
@@ -18,11 +9,7 @@ import { fromLonLat } from 'ol/proj'
 import { ipcRenderer } from 'electron'
 
 import Overview from './basemap/Overview'
-import Url from './basemap/Url'
-import XYZOptions from './basemap/XYZOptions'
-import WMTSOptions from './basemap/WMTSOptions'
-import Name from './basemap/Name'
-
+import SourceDescriptorDetails from './basemap/SourceDescriptorDetails'
 
 import { useTranslation } from 'react-i18next'
 
@@ -89,172 +76,7 @@ const useStyles = makeStyles(theme => ({
 
 
 
-const DescriptorDetails = props => {
-  const { classes, t, selectedDescriptor } = props
-  const { onCancel, onSave } = props
 
-  /* Metadata are a shallow copy without the embedded 'options' object */
-  const [metadata, setMetadata] = React.useState({ ...selectedDescriptor })
-  /* Options are required in order to create an instance of a OpenLayers source */
-  const [options, setOptions] = React.useState({ ...selectedDescriptor.options })
-
-  const [stepIndex, setStepIndex] = React.useState(0)
-  const [allowNextStep, setAllowNextStep] = React.useState(true)
-
-  const nextStep = () => setStepIndex(stepIndex => stepIndex + 1)
-  const previousStep = () => setStepIndex(stepIndex => stepIndex - 1)
-
-  const mergeOptions = (key, value) => {
-    const shadow = { ...options }
-    shadow[key] = value
-    setOptions(shadow)
-  }
-
-  const mergeMetadata = (key, value) => {
-    const shadow = { ...metadata }
-    shadow[key] = value
-    setMetadata(shadow)
-  }
-
-  const handleSaveButtonClick = () => {
-    const descriptor = { ...metadata }
-    descriptor.options = { ...options }
-    /* call the parent's onSave function */
-    onSave(descriptor)
-  }
-
-  const steps = [
-    t('basemapManagement.stepUrl'),
-    t('basemapManagement.stepOptions'),
-    t('basemapManagement.stepPreview'),
-    t('basemapManagement.stepFinalize')
-  ]
-
-  const Options = props => {
-    switch (metadata.type) {
-      case 'XYZ': return <XYZOptions options={props.options}/>
-      case 'WMTS': return <WMTSOptions options={props.options}/>
-      default: return <Typography>{t('basemapManagement.unknownSource')}</Typography>
-    }
-  }
-  Options.propTypes = { options: PropTypes.object }
-
-  const Verify = props => {
-    const { metadata, options, onValidation } = props
-    React.useEffect(() => {
-      const descriptor = { ...metadata }
-      descriptor.options = { ...options }
-      setBasemap(descriptor)
-      // visual verification is done by the user and always returns true
-      onValidation(true)
-    })
-    return <Typography>{t('basemapManagement.usePreview')}</Typography>
-  }
-  Verify.propTypes = {
-    metadata: PropTypes.object,
-    options: PropTypes.object,
-    onValidation: PropTypes.func
-  }
-
-  const getStepContent = index => {
-    switch (index) {
-      case 0:
-        return <Url
-          classes={classes}
-          // eslint-disable-next-line react/prop-types
-          url={options.url}
-          onValidation={setAllowNextStep}
-          onTypePrediction={predictedType => mergeMetadata('type', predictedType)}
-          onUrlReady={url => mergeOptions('url', url)}
-        />
-      case 1:
-      {
-        // eslint-disable-next-line react/prop-types
-        switch (metadata.type) {
-          case 'XYZ': return <XYZOptions options={options} merge={mergeOptions}/>
-          case 'WMTS': return <WMTSOptions options={options} merge={mergeOptions} onValidation={setAllowNextStep}/>
-          default: return <div>UNKNOWN SOURCE TYPE</div>
-        }
-      }
-      case 2:
-        return <Verify
-          metadata={metadata}
-          options={options}
-          onValidation={setAllowNextStep}
-        />
-      case 3:
-        return <Name
-          classes={classes}
-          // eslint-disable-next-line react/prop-types
-          name={metadata.name}
-          onValidation={setAllowNextStep}
-          onNameReady={name => mergeMetadata('name', name)}
-        />
-      default: return <div>UNKNOWN STEP</div>
-    }
-  }
-
-  return (
-    <>
-      <div>
-        <Button id="cancel" variant="contained"
-          startIcon={<CloseIcon />}
-          onClick={onCancel} >
-          {t('basemapManagement.cancel')}
-        </Button>
-      </div>
-      <div>
-        <Stepper activeStep={stepIndex} orientation="vertical">
-          { steps.map(label => {
-            return (
-              <Step key={label}>
-                <StepLabel>{label}</StepLabel>
-                <StepContent>
-                  { getStepContent(stepIndex) }
-                  <div className={classes.actions}>
-
-                    { stepIndex < (steps.length - 1)
-                      ? (<Button id="nextStep" variant="contained" color="primary" className={classes.actionButton}
-                        endIcon={<ArrowForwardIosIcon />}
-                        onClick={nextStep}
-                        disabled={!allowNextStep}
-                      >
-                        {t('basemapManagement.nextStep')}
-                      </Button>)
-                      : (<Button id="save" variant="contained" color="primary" className={classes.actionButton}
-                        startIcon={<SaveIcon />}
-                        onClick={handleSaveButtonClick}
-                        disabled={!allowNextStep}
-                      >
-                        {t('basemapManagement.save')}
-                      </Button>)
-                    }
-                    <Button id="previousStep" variant="contained" className={classes.actionButton}
-                      startIcon={<ArrowBackIosIcon />}
-                      onClick={previousStep}
-                      disabled={stepIndex === 0}
-                    >
-                      {t('basemapManagement.previousStep')}
-                    </Button>
-                  </div>
-                </StepContent>
-              </Step>
-            )
-          })
-          }
-        </Stepper>
-      </div>
-    </>
-  )
-}
-DescriptorDetails.propTypes = {
-  classes: PropTypes.object,
-  t: PropTypes.func,
-  selectedDescriptor: PropTypes.object,
-  onCancel: PropTypes.func,
-  onSave: PropTypes.func,
-  onVerify: PropTypes.func
-}
 
 
 
@@ -328,7 +150,7 @@ const BasemapManagement = props => {
       </div>
       <div className={classes.sources}>
         { isEditing
-          ? <DescriptorDetails classes={classes} t={t}
+          ? <SourceDescriptorDetails classes={classes} t={t}
             selectedDescriptor={selectedDescriptor}
             onSave={handleEditSave}
             onCancel={handleEditCancel}
